landmark-recognition/src/requirements.txt

torchvision
vit-pytorch
pillow
